Update UUIDs every month (#175)
This has not been tested at all, but I have literally just copy and pasted it from something that I _know_ that does work, so it should be fine I guess? Granted, I did change the command to execute the script so there is a very high chance that that single change has doomed the whole website. Looks like we won't know for sure until the 1st of November.